refactor(types): relocate autocomplete provider config to kilocode
